allow "perfect" matches to be regarded as isInformative()
Normal code requires a delta between best and second-best scores. In base format, there are cases where the two top scores come up as 0.0 (perfect).

This needs to be further explained, as well as the fact that in flow format, the score comes up at times position.

In base format, there are cases where the QUAL is ! - which gets translated to a prob of 1.0, which leads to a column in the matrix with two values of 1.0, which leads to accepting an insert as perfect.

In flow format, there are actually entries larger than 1.0 ... why?